refactor(reactstrap-validation-date): add types folder and run formatter
