+++ packages/reactstrap-validation-date/index.d.ts
@@ -1,7 +1,4 @@
-import AvDate from './typings/AvDate';
-import AvDateField from './typings/AvDateField';
-import AvDateRange from './typings/AvDateRange';
-import AvDateRangeField from './typings/AvDateRangeField';
-
-export default AvDate;
-export { AvDate, AvDateField, AvDateRange, AvDateRangeField };
+export { default, default as AvDate } from './types/AvDate';
+export { default as AvDateField } from './types/AvDateField';
+export { default as AvDateRange } from './types/AvDateRange';
+export { default as AvDateRangeField } from './types/AvDateRangeField';
