+++ packages/reactstrap-validation-date/package.json
@@ -3,6 +3,7 @@
   "version": "4.1.0",
   "description": "Wrapper for react-date-range to work with availity-reactstrap-validation",
   "main": "index.js",
+  "types": "index.d.ts",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -23,11 +24,9 @@
   },
   "homepage": "https://github.com/Availity/availity-react/tree/master/packages/reactstrap-validation-date#readme",
   "devDependencies": {
-    "@types/react-dates": "^21.8.2",
     "availity-reactstrap-validation": "^2.7.1",
     "date-fns": "^2.16.1",
     "react": "^17.0.2",
-    "react-date-range": "^1.1.3",
     "react-dom": "^17.0.2",
     "reactstrap": "^8.9.0"
   },
@@ -39,6 +38,7 @@
   },
   "dependencies": {
     "@availity/icon": "^0.8.0",
+    "@types/react-dates": "^21.8.2",
     "classnames": "^2.3.1",
     "lodash": "^4.17.21",
     "moment": "^2.29.1",
