+++ packages/reactstrap-validation-date/types/AvDate.d.ts
@@ -1,0 +1,11 @@
+import { AvInput } from './AvInput';
+import { SingleDatePickerShape } from 'react-dates';
+
+export interface AvDateProps extends AvInput, SingleDatePickerShape {
+  id: any;
+  datepicker?: boolean;
+}
+
+declare const AvDate: React.ComponentType<AvDateProps>;
+
+export default AvDate;
